Backend/query_api/requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform
pydantic
python-dotenv
google-cloud-firestore
python-multipart
google-auth
google-auth-oauthlib
firebase-admin
python-jose[cryptography]
slowapi==0.1.9
requests
